# Conflicts:
#	src/main/java/csc/repository/InvoiceRepository.java
#	src/main/java/csc/repository/UserRepository.java

+++ src/main/java/csc/service/TypeInvoiceService.java
@@ -3,5 +3,5 @@ package csc.service;
 import csc.models.TypeInvoice;
 
 public interface TypeInvoiceService {
-	TypeInvoice findById(long id);
+	TypeInvoice findById(Integer id);
 }
